unify into a single interface
